requirements.txt
yfinance
pandas
numpy
matplotlib.pyplot
scikit-learn
logging
timesfm